requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
blessed==1.20.0
cached-property==1.5.2
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorful==0.5.5
configparser==5.3.0
dataclasses==0.6
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.9.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.1
googleapis-common-protos==1.58.0
gpustat==1.0.0
grpcio==1.51.3
idna==3.4
jsonschema==4.17.3
ljpeg==3.7.1
minio==7.1.13
msgpack==1.0.4
multidict==6.0.4
numpy==1.24.2
nvidia-ml-py==11.495.46
opencensus==0.11.1
opencensus-context==0.1.3
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.0.0
prometheus-client==0.16.0
protobuf==4.22.0
psutil==5.9.4
py-spy==0.3.14
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
ray==2.3.0
requests==2.28.2
requests-cache==1.0.1
rsa==4.9
six==1.16.0
smart-open==6.3.0
tqdm==4.64.1
typing_extensions==4.5.0
url-normalize==1.4.3
urllib3==1.26.14
virtualenv==20.19.0
wcwidth==0.2.6
yarl==1.8.2
